Straighten out API v1 endpoints
Previously the endpoints were unversioned and partially inconsistent.

Related: #617
