+++ fmn/api/handlers/__init__.py
@@ -1,0 +1,2 @@
+from . import misc  # noqa: F401
+from . import users  # noqa: F401
