+++ frontend/src/api/rules.ts
@@ -7,7 +7,7 @@ import type { Rule } from "./types";
 // Get all rules
 export const useRulesQuery = () => {
   const userStore = useUserStore();
-  const url = `/user/${userStore.username}/rules/`;
+  const url = `/api/v1/users/${userStore.username}/rules`;
   return useQuery(url, apiGet as QueryFunction<Rule[]>);
 };
 
@@ -17,10 +17,12 @@ export const useAddRuleMutation = () => {
   const client = useQueryClient();
 
   return useMutation<Rule, unknown, Rule>(
-    (data) => apiPost(`/user/${userStore.username}/rules/`, data),
+    (data) => apiPost(`/api/v1/users/${userStore.username}/rules`, data),
     {
       onSuccess: async () => {
-        await client.invalidateQueries([`/user/${userStore.username}/rules/`]);
+        await client.invalidateQueries([
+          `/api/v1/users/${userStore.username}/rules`,
+        ]);
       },
     }
   );
@@ -30,11 +32,13 @@ export const useAddRuleMutation = () => {
 export const useEditRuleMutation = (id: number) => {
   const userStore = useUserStore();
   const client = useQueryClient();
-  const url = `/user/${userStore.username}/rules/${id}`;
+  const url = `/api/v1/users/${userStore.username}/rules/${id}`;
   return useMutation<Rule, unknown, Rule>((data) => apiPut(url, data), {
     onSuccess: async () => {
       await client.invalidateQueries([url]);
-      await client.invalidateQueries([`/user/${userStore.username}/rules/`]);
+      await client.invalidateQueries([
+        `/api/v1/users/${userStore.username}/rules`,
+      ]);
     },
   });
 };
@@ -44,10 +48,10 @@ export const useDeleteRuleMutation = () => {
   const userStore = useUserStore();
   const client = useQueryClient();
   return useMutation<void, unknown, number>(
-    (id) => apiDelete(`/user/${userStore.username}/rules/${id}`),
+    (id) => apiDelete(`/api/v1/users/${userStore.username}/rules/${id}`),
     {
       onSuccess: async () => {
-        const url = `/user/${userStore.username}/rules/`;
+        const url = `/api/v1/users/${userStore.username}/rules`;
         await client.invalidateQueries(url, { refetchActive: true });
       },
     }
